requirements.txt
absl-py==0.6.1
astor==0.7.1
atari-py==0.1.1
certifi==2018.10.15
chardet==3.0.4
future==0.17.1
gast==0.2.0
grpcio==1.16.1
gym==0.10.9
h5py==2.8.0
idna==2.7
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.4
protobuf==3.6.1
pyglet==1.3.2
requests==2.20.1
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
